# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Stock Management",
        "description": "API for managing stock items, categories, and sales",
        "contact": {
            "email": "inolatse@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/article/create": {
            "post": {
                "tags": [
                    "Article"
                ],
                "summary": "Create a new article",
                "operationId": "502177eb5601f1141ed7b0fee507aad9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "priceInit",
                                    "available_quantity",
                                    "categorie_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Article"
                                    },
                                    "priceInit": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "100.00"
                                    },
                                    "available_quantity": {
                                        "type": "integer",
                                        "example": "50"
                                    },
                                    "categorie_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/article/{id}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Get article by ID",
                "operationId": "81b4bdb0efd68cfc183904aaf3dc4001",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Article"
                ],
                "summary": "Update article by ID",
                "operationId": "607974371c23919a07f5d2b09049ecf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "priceInit",
                                    "available_quantity",
                                    "categorie_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Article Name"
                                    },
                                    "priceInit": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "120.00"
                                    },
                                    "available_quantity": {
                                        "type": "integer",
                                        "example": "70"
                                    },
                                    "categorie_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Article"
                ],
                "summary": "Delete article by ID",
                "operationId": "9ceaa4b9ca677c5d4f9d448d1ce10bf3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article was deleted successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Article": {
                "required": [
                    "name",
                    "priceInit",
                    "priceSell",
                    "available_quantity",
                    "categorie_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "priceInit": {
                        "type": "number",
                        "format": "float"
                    },
                    "priceSell": {
                        "type": "number",
                        "format": "float"
                    },
                    "available_quantity": {
                        "type": "integer"
                    },
                    "categorie_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}